I have fixed seg. fault

+++ include/Actor.h
@@ -9,9 +9,9 @@ typedef struct Actor {
 	uint8_t stamina;
 } Actor;
 
-Actor* newActor(unsigned short x, unsigned short y, size_t spriteIndex, uint8_t health, uint8_t stamina);
+Actor* newActor(unsigned int x, unsigned int y, size_t spriteIndex, uint8_t health, uint8_t stamina);
 
 void destructActor(Actor* actor);
 
-void moveActor(Actor* actor, unsigned short nx, unsigned short ny);
+void moveActor(Actor* actor, unsigned int nx, unsigned int ny);
