+++ include/Object.h
@@ -1,12 +1,12 @@
 #include <stdlib.h>
 
 typedef struct Object{
-	unsigned short x;
-	unsigned short y;
+	unsigned int x;
+	unsigned int y;
 
 	size_t spriteIndex;
 } Object;
 
-Object* newObject(unsigned short x, unsigned short y, size_t spriteIndex);
+Object* newObject(unsigned int x, unsigned int y, size_t spriteIndex);
 
 void destructObject(Object* object);
