+++ src/Actor.c
@@ -1,6 +1,6 @@
 #include "../include/Actor.h"
 
-Actor* newActor(unsigned short x, unsigned short y, size_t spriteIndex, uint8_t health, uint8_t stamina){
+Actor* newActor(unsigned int x, unsigned int y, size_t spriteIndex, uint8_t health, uint8_t stamina){
         Actor* tActor = malloc(sizeof(Actor));
 	
 	Object* tObject = newObject(x, y, spriteIndex);
@@ -16,7 +16,7 @@ void destructActor(Actor* actor){
 	free(actor);
 }
 
-void moveActor(Actor* actor, unsigned short nx, unsigned short ny){
+void moveActor(Actor* actor, unsigned int nx, unsigned int ny){
         actor->object->x = nx;
         actor->object->y = ny;
 }
