+++ src/Player.c
@@ -1,7 +1,7 @@
 #include "../include/Player.h"
 
 
-Player* initPlayer(unsigned short x, unsigned short y, size_t spriteIndex, uint8_t health, uint8_t stamina, uint8_t bullets){
+Player* initPlayer(unsigned int x, unsigned int y, size_t spriteIndex, uint8_t health, uint8_t stamina, uint8_t bullets){
 	Player* tPlayer = malloc(sizeof(Player));
 	Actor* tActor = newActor(x, y, spriteIndex, health, stamina);
